pyproject.toml
[tool.poetry]
name = "cray-freelas-bot"
version = "0.1.0"
description = ""
authors = ["riguima <richard.alexsander.guima@gmail.com>"]
readme = "README.md"
packages = [{include = "cray_freelas_bot"}]

[tool.poetry.dependencies]
python = "^3.9"
selenium = "^4.9.1"
webdriver-manager = "^3.8.6"
jedi = "^0.18.2"
flake8 = "^6.0.0"
python-slugify = "^8.0.1"
httpx = "^0.24.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
